# mypaintings/target/classes/api-spec.yaml
openapi: 3.0.0
info:
  title: MyP@intings
  version: 1.0.0
  description: RESTful API dla aplikacji myp@intings
...
paths:
...
  /artysta/mojedziela:
    get:
      tags:
        - Artysta
      summary: Pobierz listę dzieł artysty
      description: Endpoint zwraca listę dzieł artysty wraz z linkami do zdjęć.
      parameters:
        - name: Authorization
          in: header
          description: Token JWT w formacie "Bearer {token}"
          required: true
          schema:
            type: string
            example: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: Sukces
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response_artysta_mojedziela'
        '401':
          description: Brak autoryzacji
        '403':
          description: Brak dostępu
  /artysta/sales:
    get:
      tags:
        - Artysta
      summary: Pobierz listę sprzedanych dzieł artysty
      description: Endpoint zwraca listę sprzedanych dzieł artysty wraz z linkami do zdjęć.
      parameters:
        - name: Authorization
          in: header
          description: Token JWT w formacie "Bearer {token}"
          required: true
          schema:
            type: string
            example: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: Sukces
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response_artysta_sales'
        '401':
          description: Brak autoryzacji
        '403':
          description: Brak dostępu
...
  /artysta/dzielo:
    post:
      tags:
        - Artysta
      summary: Dodaj nowe dzieło sztuki
      description: Endpoint pozwala artyście dodać nowe dzieło sztuki do systemu.
      parameters:
        - name: Authorization
          in: header
          description: Token JWT w formacie "Bearer {token}"
          required: true
          schema:
            type: string
            example: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/request_artysta_dzielo'
      responses:
        '201':
          description: Dzieło sztuki zostało dodane pomyślnie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response_artysta_dzielo'
        '400':
          description: Nieprawidłowe dane wejściowe
        '401':
          description: Brak autoryzacji
...
  /artysta/stronahotelu/{id_hotelu}:
    get:
      tags:
        - Artysta
      summary: Pobierz link do strony hotelu
      description: Endpoint zwraca link do strony hotelu o podanym ID.
      parameters:
        - name: id_hotelu
          in: path
          description: ID hotelu
          required: true
          schema:
            type: integer
            example: 456
      responses:
        '200':
          description: Sukces
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response_artysta_stronahotelu'
        '404':
          description: Hotel nie znaleziony lub brak strony
...
  /artysta/dzielo/{image_id}:
    delete:
      tags:
        - Artysta
      summary: Usuń dzieło artysty
      description: Endpoint pozwala artyście usunąć swoje dzieło z systemu.
      parameters:
        - name: Authorization
          in: header
          description: Token JWT w formacie "Bearer {token}"
          required: true
          schema:
            type: string
            example: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        - name: image_id
          in: path
          description: ID dzieła do usunięcia
          required: true
          schema:
            type: integer
            example: 123
      responses:
        '200':
          description: Dzieło zostało usunięte pomyślnie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response_artysta_dzielo_delete'
        '401':
          description: Brak autoryzacji lub użytkownik nie jest autorem dzieła
        '404':
          description: Dzieło nie znalezione
        '409':
          description: Nie można usunąć dzieła (np. jest już zarezerwowane lub sprzedane)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response_artysta_dzielo_delete_conflict'
...
components:
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    response_artysta_mojedziela:
      type: object
      properties:
        dziela:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                example: 123
              date_of_post:
                type: string
                format: date
                example: "2024-11-11"
              price:
                type: string
                example: "$600"
              hotel:
                type: string
                example: "@NY_Hotel_Star"
              viewers:
                type: integer
                example: 59
              likes:
                type: integer
                example: 12
              image_url:
                type: string
                example: "https://example.com/images/image.png"
    response_artysta_sales:
      type: object
      properties:
        sales:
          type: array
          items:
            type: object
            properties:
              sold_on:
                type: string
                format: date
                example: "2024-12-11"
              price:
                type: string
                example: "$650"
              hotel:
                type: string
                example: "@GoldenHotelNY"
              likes:
                type: integer
                example: 13
              image_url:
                type: string
                example: "https://example.com/images/image.png"
...
    response_artysta_dzielo:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: "Pejzaż górski"
        dimensions:
          type: string
          example: "100x150 cm"
        price:
          type: number
          example: 1200
        category:
          type: string
          example: "Malarstwo"
        medium:
          type: string
          example: "Olej na płótnie"
        style:
          type: string
          example: "Realizm"
        date:
          type: string
          format: date
          example: "2024-01-15"
        description:
          type: string
          example: "Pejzaż przedstawiający góry o zachodzie słońca."
...
    response_artysta_stronahotelu:
      type: object
      properties:
        website_url:
          type: string
          example: "https://www.example-hotel.com"
...
    response_artysta_dzielo_delete:
      type: object
      properties:
        message:
          type: string
          example: "Dzieło zostało pomyślnie usunięte."
        deleted_image_id:
          type: integer
          example: 123
    response_artysta_dzielo_delete_conflict:
      type: object
      properties:
        error:
          type: string
          example: "Nie można usunąć dzieła, ponieważ jest już zarezerwowane."
        current_status:
          type: string
          example: "RESERVED"
...
    request_artysta_dzielo:
      type: object
      properties:
        title:
          type: string
          example: "Pejzaż górski"
        dimensions:
          type: string
          example: "100x150 cm"
        price:
          type: number
          example: 1200
        category:
          type: string
          example: "Malarstwo"
        medium:
          type: string
          example: "Olej na płótnie"
        style:
          type: string
          example: "Realizm"
        date:
          type: string
          format: date
          example: "2024-01-15"
        description:
          type: string
          example: "Pejzaż przedstawiający góry o zachodzie słońca."
      required:
        - title
        - dimensions
        - price
        - category
        - medium
        - style
        - date
        - description
...